refactor: create root dir for token-metadata, separate into files
