+++ src/api/routes/tokens/tokens.ts
@@ -14,10 +14,7 @@ import {
   NonFungibleTokensMetadataList,
 } from '@stacks/stacks-blockchain-api-types';
 import { parseLimitQuery, parsePagingQueryInput } from './../../pagination';
-import {
-  isFtMetadataEnabled,
-  isNftMetadataEnabled,
-} from '../../../event-stream/tokens-contract-handler';
+import { isFtMetadataEnabled, isNftMetadataEnabled } from '../../../token-metadata/helpers';
 import { bufferToHexPrefixString, has0xPrefix, isValidPrincipal } from '../../../helpers';
 import { booleanValueForParam, isUnanchoredRequest } from '../../../api/query-helpers';
 import { decodeClarityValueToRepr } from 'stacks-encoding-native-js';
