+++ src/index.ts
@@ -15,11 +15,6 @@ import { PgDataStore } from './datastore/postgres-store';
 import { startApiServer } from './api/init';
 import { startProfilerServer } from './inspector-util';
 import { startEventServer } from './event-stream/event-server';
-import {
-  isFtMetadataEnabled,
-  isNftMetadataEnabled,
-  TokensProcessorQueue,
-} from './event-stream/tokens-contract-handler';
 import { StacksCoreRpcClient } from './core-rpc/client';
 import { createServer as createPrometheusServer } from '@promster/server';
 import { registerShutdownConfig } from './shutdown-handler';
@@ -30,6 +25,8 @@ import * as getopts from 'getopts';
 import * as fs from 'fs';
 import { injectC32addressEncodeCache } from './c32-addr-cache';
 import { exportEventsAsTsv, importEventsFromTsv } from './event-replay/event-replay';
+import { isFtMetadataEnabled, isNftMetadataEnabled } from './token-metadata/helpers';
+import { TokensProcessorQueue } from './token-metadata/tokens-processor-queue';
 
 enum StacksApiMode {
   /**
