+++ src/rosetta-helpers.ts
@@ -61,9 +61,8 @@ import { getCoreNodeEndpoint } from './core-rpc/client';
 import { getBTCAddress, poxAddressToBtcAddress } from '@stacks/stacking';
 import {
   tokenMetadataErrorMode,
-  isFtMetadataEnabled,
   TokenMetadataErrorMode,
-} from './event-stream/tokens-contract-handler';
+} from './token-metadata/tokens-contract-handler';
 import {
   ClarityTypeID,
   decodeClarityValue,
@@ -79,6 +78,7 @@ import {
   TxPayloadTokenTransfer,
   TxPayloadTypeID,
 } from 'stacks-encoding-native-js';
+import { isFtMetadataEnabled } from './token-metadata/helpers';
 
 enum CoinAction {
   CoinSpent = 'coin_spent',
